# src/Encore.Api/wwwroot/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Encore",
    "version": "0.1.0",
    "description": "Event ticketing, as a modular monolith. Five modules, and since the Payments extraction two hosts: Encore.Api serves everything here except the three /internal/payments routes, which only Encore.Payments.Api maps. Paths that carry their own `servers` entry are the ones the host serving this page does not answer.\n\n**Actions, not resources.** The seat routes are `/hold`, `/release` and `/purchase` rather than a `/holds` collection: a hold is not an entity, it is two fields on the seat row, so it gets no URI of its own. All three are idempotent, which is what makes retrying a POST safe.\n\n**Branch on `reason`, not on status.** Every refusal about the state of the world is `409` with a machine-readable `reason` and a `retriable` flag. Only an addressed thing that is not there is `404`. A request that is wrong on its own terms is `400`. The statuses carry the class of failure; `reason` carries which one.\n\n**`X-Client-Id` is a claimed identity, not authentication.** Anyone can change it and reset their own hold cap. It stands in for the Identity module, which does not exist yet, which is also why no route returns 401 or 403.\n\nThis document is hand-maintained and was verified against a running instance. `Encore.Api` holds zero NuGet packages by design, so nothing generates this from endpoint metadata; a test fails when it and the mapped routes disagree."
  },
  "servers": [
...
  "components": {
    "securitySchemes": {
      "ClientId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "A GUID identifying the calling client. NOT authentication — it is unverified and anyone may change it. Paste any GUID here to try the seat, order and payment routes."
      },
      "ServiceToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Token",
        "description": "The shared secret between Orders and Payments. This one IS authentication, and it is the only thing guarding a call that can move money. Not a customer credential — the /internal routes take no X-Client-Id at all."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details. Module refusals add `reason` and `retriable`; framework refusals (404 for an unmatched route, 405, 415, a body that will not bind) carry only the standard members.",
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" },
          "reason": { "type": "string", "description": "Machine-readable refusal code. Branch on this." },
          "retriable": { "type": "boolean", "description": "Whether the identical request could plausibly succeed later." },
          "limit": { "type": "integer", "description": "Present on hold_cap_reached and too_many_seats." },
          "traceId": { "type": "string" }
        },
        "required": ["type", "title", "status"]
      },
      "CreateVenueRequest": {
        "type": "object",
        "required": ["name", "address", "capacity"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "address": { "type": "string", "maxLength": 500 },
          "capacity": { "type": "integer", "minimum": 1 }
        }
      },
      "VenueResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "capacity": { "type": "integer" }
        }
      },
      "CreateEventRequest": {
        "type": "object",
        "required": ["venueId", "name", "startsAt", "price", "currency"],
        "properties": {
          "venueId": { "type": "string", "format": "uuid", "description": "Must already exist." },
          "name": { "type": "string", "maxLength": 200 },
          "startsAt": { "type": "string", "format": "date-time", "description": "Must carry a timezone. Stored and returned as UTC." },
          "onSaleAt": { "type": ["string", "null"], "format": "date-time", "description": "When tickets become orderable. Null means immediately. Must carry a timezone when supplied." },
          "price": { "type": "number", "description": "What one seat costs." },
          "currency": { "type": "string", "description": "ISO 4217 code.", "example": "GBP" }
        }
      },
      "EventResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "venueId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "startsAt": { "type": "string", "format": "date-time" },
          "onSaleAt": { "type": ["string", "null"], "format": "date-time" },
          "price": { "type": "number" },
          "currency": { "type": "string" }
        }
      },
...
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "orderId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["pending", "authorized", "captured", "declined", "timed_out", "voided"] },
          "amount": { "type": "number" },
          "currency": { "type": "string" },
          "gatewayReference": { "type": ["string", "null"] },
          "attemptedAt": { "type": "string", "format": "date-time" },
          "resolvedAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "BadClientId": {
        "description": "The X-Client-Id header is missing or is not a GUID. No `reason`: the request is malformed rather than refused.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
...
    },
    "parameters": {
      "EventId": { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
    }
  },
  "paths": {
...
    "/catalog/venues": {
      "post": {
        "tags": ["catalog"],
        "summary": "Creates a venue and returns it.",
        "description": "Operator-facing. Ids are generated here, not supplied.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateVenueRequest" } } } },
        "responses": {
          "201": {
            "description": "Created.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VenueResponse" } } }
          },
          "400": {
            "description": "reason: invalid_name, invalid_address, invalid_capacity.",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
          }
        }
      },
      "get": {
        "tags": ["catalog"],
        "summary": "Lists every venue.",
        "responses": {
          "200": {
            "description": "All venues.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/VenueResponse" } } } }
          }
        }
      }
    },
    "/catalog/venues/{venueId}": {
      "get": {
        "tags": ["catalog"],
        "summary": "Reads one venue.",
        "parameters": [{ "name": "venueId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "The venue.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VenueResponse" } } } },
          "404": { "description": "reason: venue_not_found.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    },
    "/catalog/events": {
      "post": {
        "tags": ["catalog"],
        "summary": "Creates an event at an existing venue and returns it.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateEventRequest" } } } },
        "responses": {
          "201": { "description": "Created.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventResponse" } } } },
          "400": {
            "description": "reason: invalid_name, invalid_price, invalid_currency, ambiguous_timestamp. The last one is a startsAt or onSaleAt with no timezone.",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
          },
          "409": { "description": "reason: venue_not_found. Create the venue before the event.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      },
      "get": {
        "tags": ["catalog"],
        "summary": "Lists every event, soonest first.",
        "responses": {
          "200": { "description": "All events.", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/EventResponse" } } } } }
        }
      }
    },
    "/catalog/events/{eventId}": {
      "get": {
        "tags": ["catalog"],
        "summary": "Reads one event.",
        "parameters": [{ "$ref": "#/components/parameters/EventId" }],
        "responses": {
          "200": { "description": "The event.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventResponse" } } } },
          "404": { "description": "reason: event_not_found.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    },
...
    "/payments/{paymentId}": {
      "get": {
        "tags": ["payments"],
        "summary": "Reads back what happened to a payment.",
        "security": [{ "ClientId": [] }],
        "parameters": [{ "name": "paymentId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "The attempt.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentResponse" } } } },
          "400": { "$ref": "#/components/responses/BadClientId" },
          "404": { "description": "reason: payment_not_found.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    },
    "/payments": {
      "get": {
        "tags": ["payments"],
        "summary": "Lists the attempts against one order.",
        "description": "Read-only on purpose: Orders drives payment, because Orders is the thing that knows what is owed. There is no POST here.",
        "security": [{ "ClientId": [] }],
        "parameters": [{ "name": "orderId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Every attempt against that order, including superseded ones.", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PaymentResponse" } } } } },
          "400": { "$ref": "#/components/responses/BadClientId" }
        }
      }
    },
...
  }
}